update build pipeline & refine gpu stream

+++ superbench/benchmarks/micro_benchmarks/gpu_stream/gpu_stream_kernels.cu
@@ -137,7 +137,7 @@ __global__ void AddKernel(double *tgt, const double *src_a, const double *src_b)
  * @param[in] src_b The second source array containing the second set of operands to be multiplied by the scalar.
  * @param[in] scalar The scalar value used in the multiply/add operation.
  */
-__global__ void TriadKernel(double *tgt, const double *src_a, const double *src_b, const long scalar) {
+__global__ void TriadKernel(double *tgt, const double *src_a, const double *src_b, const double scalar) {
     uint64_t index = blockIdx.x * blockDim.x * kNumLoopUnrollAlias + threadIdx.x;
     double val_a[kNumLoopUnrollAlias];
     double val_b[kNumLoopUnrollAlias];
